rusk: refactor CLI commands
- Move `recovery-keys` to `recovery keys`
- Move `recovery-state` to `recovery state`
- Add `chain revert`
