+++ rusk/src/bin/args.rs
@@ -4,8 +4,7 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-#[cfg(any(feature = "recovery-state", feature = "recovery-keys"))]
-mod command;
+pub(crate) mod command;
 #[cfg(feature = "recovery-state")]
 mod state;
 
@@ -97,8 +96,7 @@ pub struct Args {
     /// Kadcast network id
     pub kadcast_network_id: Option<u8>,
 
-    #[cfg(any(feature = "recovery-state", feature = "recovery-keys"))]
-    /// Command
+    /// Utility commands
     #[clap(subcommand)]
     pub command: Option<command::Command>,
 }
